RAG_Project/requirements.txt
# Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0

# Vector store
chromadb>=0.5.0

# Embeddings & LLM
openai>=1.40.0

# Dataset
datasets>=2.20.0
huggingface-hub>=0.24.0

# Phase 2: Hybrid search + reranking
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# Phase 3: Evaluation
ragas>=0.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.26.0
pandas>=2.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Testing
pytest>=8.0.0
